structuring application
