--- src/\
@@ -1,82 +1,0 @@
-mod component;
-mod components;
-mod database;
-mod home;
-
-use crate::component::Component;
-use crate::components::{Components, CurrentComponent};
-use iced::{
-    widget::{button, column, container, row, text, Column},
-    Element, Task,
-};
-
-#[derive(Debug, Clone)]
-pub enum Message {
-    InitializeComponents(Components),
-    InitializeHomeComponent,
-}
-
-struct Crm {
-    current_component: CurrentComponent,
-    components: Option<Components>,
-}
-
-impl Crm {
-    fn new() -> (Self, Task<Message>) {
-        (
-            Self {
-                current_component: CurrentComponent::Home,
-                components: None,
-            },
-            Task::perform(Components::new(), |components| {
-                Message::InitializeComponents(components);
-                Message::InitializeHomeComponent
-            }),
-        )
-    }
-    fn title(&self) -> String {
-        String::from("CRM")
-    }
-    fn view(&self) -> Element<Message> {
-        if self.components.is_none() {
-            column![container("loading")].into()
-        } else {
-            let components = self.components.clone().unwrap();
-            match self.current_component {
-                CurrentComponent::Home => {
-                    let home_component = components.home;
-                    if !home_component.tables.is_none() {
-                        column!(home_component
-                            .tables
-                            .unwrap()
-                            .iter()
-                            .map(|table| table))
-                        .into()
-                    } else {
-                        column!(text("loading")).into()
-                    }
-                }
-            }
-        }
-    }
-    fn update(&mut self, message: Message) -> Task<Message> {
-        match message {
-            Message::InitializeComponents(components) => {
-                self.components = Some(components);
-                Task::none()
-            }
-            Message::InitializeHomeComponent => {
-                let mut home_component = self.components.clone().unwrap().home;
-                Task::perform(
-                    async move { home_component.initialize_component().await },
-                    |_| (),
-                );
-                Task::none()
-            }
-        }
-    }
-}
-
-fn main() -> iced::Result {
-    iced::application(Crm::title, Crm::update, Crm::view).run_with(Crm::new)
-}
