--- src/component.rs
@@ -1,3 +1,0 @@
-pub trait Component {
-    async fn initialize_component(&mut self) {}
-}
